requirements.txt
gunicorn==20.0.2
Django==2.2.7
django-autoslug==1.9.6
django-image-optimizer==1.0.0
django-optimized-image==0.2.0
Pillow==6.2.1
pytz==2019.3
requests==2.22.0
tinify==1.5.1
traitlets==4.3.3
urllib3==1.25.7
whitenoise==4.1.4
